interface_msg/requirements.txt
mysql-connector
sanic
sanic_httpauth